el: normalize with mdbook-i18n-helpers 0.2.2 (#1113)
Before, po/el.po had these statistics:

227 translated messages, 846 fuzzy translations, 708 untranslated
messages.

Afterwards, the statistics for po/el.po is:

236 translated messages, 992 fuzzy translations, 1239 untranslated
messages.

The number of translated messages changed from 13% to 10%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.