bn: normalize with mdbook-i18n-helpers 0.2.2 (#1112)
Before, po/bn.po had these statistics:

328 translated messages, 35 fuzzy translations, 1418 untranslated
messages.

Afterwards, the statistics for po/bn.po is:

351 translated messages, 42 fuzzy translations, 2074 untranslated
messages.

The number of translated messages changed from 18% to 14%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.