da: normalize with mdbook-i18n-helpers 0.2.2 (#1111)
Before, po/da.po had these statistics:

279 translated messages, 26 fuzzy translations, 1476 untranslated
messages.

Afterwards, the statistics for po/da.po is:

244 translated messages, 22 fuzzy translations, 2201 untranslated
messages.

The number of translated messages changed from 16% to 10%.

With this change, it becomes important to use the latest version of
mdbook-i18n-helpers when viewing the translation locally. To update to
the latest version, run

    cargo install mdbook-i18n-helpers

You will now be able to serve the translation locally.

Part of #330.