Only save CI cache on `main` branch (#1962)
When a PR restores the cache, it will find the most recent cache from
the `main` branch. This cache should be mostly up to date.

We are constantly at the 10 GB limit for our CI caches, so removing the
extra caches from the many PR branches should ensure that we always have
space for the `main` branch caches.