zh-CN: glossary translation (#1992)
The po file has been refreshed. Translation range: L22164-L22764.

https://github.com/google/comprehensive-rust/issues/324

Translations of terms refer to the Chinese version of the TRPL and
https://github.com/rust-lang-cn/english-chinese-glossary-of-rust/blob/master/rust-glossary.md

Note: Since italics are not used in Chinese to indicate quotation or
emphasisin Chinese, English italics have been translated into Chinese
bold, and the original English text of terms is marked in full-width
parentheses.